Add projectiles over the network

+++ MultiplayerShooter.Library/PlayerData.cs
@@ -2,7 +2,7 @@
 {
     public class PlayerData
     {
-        public long Id { get; set; }
+        public byte Id { get; set; }
         public string Username { get; set; }
         public int PositionX { get; set; }
         public int PositionY { get; set; }
